+++ Tests/LinuxMain.swift
@@ -1433,7 +1433,9 @@ extension SortedFirstLastRuleTests {
 
 extension SortedImportsRuleTests {
     static var allTests: [(String, (SortedImportsRuleTests) -> () throws -> Void)] = [
-        ("testWithDefaultConfiguration", testWithDefaultConfiguration)
+        ("testWithDefaultConfiguration", testWithDefaultConfiguration),
+        ("testWithTestableImportsTop", testWithTestableImportsTop),
+        ("testWithTestableImportsBottom", testWithTestableImportsBottom)
     ]
 }